+++ src/ExchangeRateProvider/PdoProvider.php
@@ -55,10 +55,11 @@ final class PdoProvider implements ExchangeRateProvider
         private readonly PDO $pdo,
         private readonly string $tableName,
         private readonly string $exchangeRateColumnName,
-        private readonly ?string $sourceCurrencyCode,
+        private readonly string|int|null $sourceCurrencyCode,
         private readonly ?string $sourceCurrencyColumnName,
-        private readonly ?string $targetCurrencyCode,
+        private readonly string|int|null $targetCurrencyCode,
         private readonly ?string $targetCurrencyColumnName,
+        private readonly bool $useNumericCurrencyCode,
         private readonly ?SqlCondition $staticCondition,
         private array $dimensionBindings,
         private readonly ?string $orderBy,
@@ -90,6 +91,7 @@ final class PdoProvider implements ExchangeRateProvider
             sourceCurrencyColumnName: $builder->getSourceCurrencyColumnName(),
             targetCurrencyCode: $builder->getTargetCurrencyCode(),
             targetCurrencyColumnName: $builder->getTargetCurrencyColumnName(),
+            useNumericCurrencyCode: $builder->getUseNumericCurrencyCode(),
             staticCondition: $builder->getStaticCondition(),
             dimensionBindings: $builder->getDimensionBindings(),
             orderBy: $builder->getOrderBy(),
@@ -103,8 +105,16 @@ final class PdoProvider implements ExchangeRateProvider
             return BigInteger::one();
         }
 
-        $sourceCurrencyCode = $sourceCurrency->getCurrencyCode();
-        $targetCurrencyCode = $targetCurrency->getCurrencyCode();
+        $sourceCurrencyCode = $this->useNumericCurrencyCode
+            ? $sourceCurrency->getNumericCode()
+            : $sourceCurrency->getCurrencyCode();
+        $targetCurrencyCode = $this->useNumericCurrencyCode
+            ? $targetCurrency->getNumericCode()
+            : $targetCurrency->getCurrencyCode();
+
+        if ($sourceCurrencyCode === null || $targetCurrencyCode === null) {
+            return null;
+        }
 
         $sqlConditions = [];